adjust lvs separator to be ';'
previously holland.lib.lvm used ',', which breaks on any fields
with an embedded comma. This became a probably when a user
passed multiple --addtag options, and a later lvs lookup
may not parse the vg_name + lv_tags properly, but is also
an existing potential issue if a volume (or volume group)
had specified multiple tags.